// src/app/shared/enums/itemDetailTypeNo.enum.ts
export enum ItemDetailType {
    Cultivar = 1,
    Ratings = 2,
    Type = 3,
    Origin = 4,
    Appellation = 5,
    Region = 6,
    TastingNoteTags = 7,
    FoodPairingTags = 8,
    Vintage = 9,
    BottleSize = 10,
    TastingNotes = 11,
    Recipe = 12,
    InTheVineyard = 13,
    InTheCellar = 14,
    Ageing = 15,
    AdditionalInformation = 16,
    Alcohol = 17,
    Sugar = 18,
    MaturationProtentialFrom = 19,
    MaturationProtentialTo = 20,
    Country = 21,
    FoodPairing = 23
}